+++ ModelGenerator/regenerate.cpp
@@ -9,7 +9,7 @@
 using namespace std;
 
 struct ppp{
-    float x;float y;float z;
+    double x;double y;double z;
 };
 
 struct cmp{
@@ -23,37 +23,37 @@ const int N=10000;
 FILE* input;
 FILE* output;
 int n,m;
-float vol,elas,dens;
-float x[N][3],y[N][3],z[N][3],vx[N],vy[N],vz[N],texx[N][3],texy[N][3];
+double vol,elas,dens;
+double x[N][3],y[N][3],z[N][3],vx[N],vy[N],vz[N],texx[N][3],texy[N][3];
 set<ppp,cmp> a;
 
 // calculate normal vector
-void normv(int loca,float x,float y,float z,float xx,float yy,float zz){
+void normv(int loca,double x,double y,double z,double xx,double yy,double zz){
     vx[loca]=y*zz-z*yy;
     vy[loca]=-x*zz+z*xx;
     vz[loca]=x*yy-y*xx;
 }
 
 // dot calculate
-float dot(float x,float y,float z,int loca){
+double dot(double x,double y,double z,int loca){
     return x*vx[loca]+y*vy[loca]+z*vz[loca];
 }
 
 // square calculate
-double sqr(float a){return a*a;}
+double sqr(double a){return a*a;}
 
 // dist calculate
-double dist(float a,float b,float c,float x,float y,float z){
+double dist(double a,double b,double c,double x,double y,double z){
     return sqrt(sqr(a-x)+sqr(b-y)+sqr(c-z));
 }
 
 // calculate part of inertia
-float calc(float ax,float ay,float bx,float by,float cx,float cy){
+double calc(double ax,double ay,double bx,double by,double cx,double cy){
     return 2*(ax*ay+bx*by+cx*cy)+(ax*by+ay*bx+ax*cy+ay*cx+bx*cy+by*cx);
 }
 
 void prt(int i,int j){
-    fprintf(output,"%.10f %.10f %.10f %.10f %.10f ",x[i][j],y[i][j],z[i][j],texx[i][j],texy[i][j]);
+    fprintf(output,"%.20lf %.20lf %.20lf %.20lf %.20lf ",x[i][j],y[i][j],z[i][j],texx[i][j],texy[i][j]);
 }
 
 void trans(){
@@ -62,12 +62,12 @@ void trans(){
     fscanf(input,"%d",&n);
     for (int i=0;i<n;++i){
         for (int j=0;j<3;++j){
-            fscanf(input,"%f%f%f%f%f",x[i]+j,y[i]+j,z[i]+j,texx[i]+j,texy[i]+j);
+            fscanf(input,"%lf%lf%lf%lf%lf",x[i]+j,y[i]+j,z[i]+j,texx[i]+j,texy[i]+j);
             maxdist=max(maxdist,dist(x[i][j],y[i][j],z[i][j],0,0,0));
         }
         normv(i,x[i][1]-x[i][0],y[i][1]-y[i][0],z[i][1]-z[i][0],
                 x[i][2]-x[i][1],y[i][2]-y[i][1],z[i][2]-z[i][1]);
-        float tv=dot(x[i][0],y[i][0],z[i][0],i);
+        double tv=dot(x[i][0],y[i][0],z[i][0],i);
         vol+=tv/6;
         ixy+=tv/120*calc(x[i][0],y[i][0],x[i][1],y[i][1],x[i][2],y[i][2]);
         ixz+=tv/120*calc(x[i][0],z[i][0],x[i][1],z[i][1],x[i][2],z[i][2]);
@@ -77,8 +77,8 @@ void trans(){
         izz+=tv/120*calc(z[i][0],z[i][0],z[i][1],z[i][1],z[i][2],z[i][2]);
     }
 
-    float dens=(rand()%100)/100.0;
-    fprintf(output,"%d %.10f %.10f %.10f %.10lf\n",n,vol,dens+10,10+(rand()%100)/100.0,maxdist);
+    double dens=(rand()%100)/100.0;
+    fprintf(output,"%d %.20lf %.20lf %.20lf %.10lf\n",n,vol,dens+10,10+(rand()%100)/100.0,maxdist);
     for (int i=0;i<n;++i){
         for (int j=0;j<3;++j){
             prt(i,j);
@@ -88,9 +88,9 @@ void trans(){
             if (a.find(now)==a.end()) {a.insert(now); fprintf(output,"1 ");}
                 else fprintf(output,"0 ");
         }
-        fprintf(output,"%.10f %.10f %.10f\n",vx[i],vy[i],vz[i]);
+        fprintf(output,"%.20lf %.20lf %.20lf\n",vx[i],vy[i],vz[i]);
     }
-    fprintf(output,"%.10f %.10f %.10f %.10f %.10f %.10f %.10f %.10f %.10f",
+    fprintf(output,"%.20lf %.20lf %.20lf %.20lf %.20lf %.20lf %.20lf %.20lf %.20lf",
                     ixx,ixy,ixz,ixy,iyy,iyz,ixz,iyz,izz);
 }
